Update README.md
The goals for this project were sourcing my own dataset and explore relationships within the data in Python. After the initial steps in the exploratory analysis I created the following hypothesis:

If the review scores rating is higher, then the review scores of the location are also higher.

I used the results that I created in Python to build a business case in Tableau.